+++ docs-src/navigation.js
@@ -1,6 +1,7 @@
 import UiAlert from './pages/UiAlert.vue';
 import UiAutocomplete from './pages/UiAutocomplete.vue';
 import UiButton from './pages/UiButton.vue';
+import UiCalendar from './pages/UiCalendar.vue';
 import UiCheckbox from './pages/UiCheckbox.vue';
 import UiCheckboxGroup from './pages/UiCheckboxGroup.vue';
 import UiCollapsible from './pages/UiCollapsible.vue';
@@ -55,6 +56,12 @@ const menu = [
                 component: UiCheckbox,
                 title: 'UiCheckbox',
                 sourceUrl: 'src/UiCheckbox.vue'
+            },
+            {
+                path: '/ui-calendar',
+                component: UiCalendar,
+                title: 'UiCalendar',
+                sourceUrl: 'src/UiCalendar.vue'
             },
             {
                 path: '/ui-checkbox-group',