+++ src/UiCalendar.vue
@@ -1,77 +1,35 @@
 <template>
     <div class="ui-calendar" :class="classes">
-        <div class="ui-calendar__header">
-            <div
-                class="ui-calendar__header-year"
-                tabindex="0"
+        <ui-calendar-controls
+            ref="controls"
+            class="ui-calendar__header"
 
-                :class="{ 'is-active': showYearPicker }"
+            :color="color"
+            :date-in-view="dateInView"
+            :lang="lang"
 
-                @click="showYearPicker = true"
-                @keydown.enter="showYearPicker = true"
-            >{{ headerYear }}</div>
-
-            <div
-                class="ui-calendar__header-details"
-                tabindex="0"
-
-                :class="{ 'is-active': !showYearPicker }"
-
-                @click="showYearPicker = false"
-                @keydown.enter="showYearPicker = false"
-            >
-                <span class="ui-calendar__header-day">{{ headerDay }}, </span>
-                <span class="ui-calendar__header-date">{{ headerDate }}</span>
-            </div>
-        </div>
+            @go-to-date="goToDate"
+        ></ui-calendar-controls>
 
         <div class="ui-calendar__body">
-            <ul class="ui-calendar__years" ref="years" v-show="showYearPicker">
-                <li
-                    class="ui-calendar__year"
-                    tabindex="0"
+            <ui-calendar-month
+                ref="month"
 
-                    :class="getYearClasses(year)"
+                :color="color"
+                :date-in-view="dateInView"
+                :lang="lang"
+                :selected="value"
+                :start-of-week="startOfWeek"
+                :square-cells="squareCells"
 
-                    @click="selectYear(year)"
-                    @keydown.enter="selectYear(year)"
-
-                    v-for="year in yearRange"
-                    v-if="!isYearOutOfRange(year)"
-                >{{ year }}</li>
-            </ul>
-
-            <div v-show="!showYearPicker">
-                <ui-calendar-controls
-                    ref="controls"
-
-                    :date-in-view="dateInView"
-                    :lang="lang"
-                    :max-date="maxDate"
-                    :min-date="minDate"
-
-                    @go-to-date="onGoToDate"
-                ></ui-calendar-controls>
-
-                <ui-calendar-month
-                    ref="month"
-
-                    :date-filter="dateFilter"
-                    :date-in-view="dateInView"
-                    :lang="lang"
-                    :max-date="maxDate"
-                    :min-date="minDate"
-                    :selected="value"
-                    :start-of-week="startOfWeek"
-
-                    @change="onMonthChange"
-                    @date-select="onDateSelect"
-                ></ui-calendar-month>
-            </div>
-
-            <div class="ui-calendar__footer" v-if="$slots.footer">
-                <slot name="footer"></slot>
-            </div>
+                @change="onMonthChange"
+                @date-select="onDateSelect"
+            >
+                <template slot-scope="props" slot="date">
+                    <slot name="date" :date="props.date" v-if="$scopedSlots.date"></slot>
+                    <template v-else>{{ props.date.getDate() }}</template>
+                </template>
+            </ui-calendar-month>
         </div>
     </div>
 </template>
@@ -81,25 +39,37 @@ import UiCalendarControls from './UiCalendarControls.vue';
 import UiCalendarMonth from './UiCalendarMonth.vue';
 
 import dateUtils from './helpers/date';
-import { scrollIntoView } from './helpers/element-scroll';
 
 export default {
     name: 'ui-calendar',
 
     props: {
-        value: Date,
-        minDate: Date,
-        maxDate: Date,
-        startOfWeek: {
-            type: Number,
-            default: 0
+        color: {
+            type: String,
+            default: 'default' // 'default', 'primary' or 'accent'
         },
+        dateFilter: Function,
         lang: {
             type: Object,
             default() {
                 return dateUtils.defaultLang;
             }
         },
+        maxDate: Date,
+        minDate: Date,
+        raised: {
+            type: Boolean,
+            default: false
+        },
+        startOfWeek: {
+            type: Number,
+            default: 0
+        },
+        squareCells: {
+            type: Boolean,
+            default: false
+        },
+        value: Date,
         yearRange: {
             type: Array,
             default() {
@@ -112,23 +82,13 @@ export default {
                         return (thisYear - 100) + index;
                     });
             }
-        },
-        dateFilter: Function,
-        color: {
-            type: String,
-            default: 'primary' // 'primary' or 'accent'
-        },
-        orientation: {
-            type: String,
-            default: 'portrait' // 'portrait' or 'landscape'
         }
     },
 
     data() {
         return {
             today: new Date(),
-            dateInView: this.getDateInRange(this.value, new Date()),
-            showYearPicker: false
+            dateInView: this.value || new Date()
         };
     },
 
@@ -136,25 +96,8 @@ export default {
         classes() {
             return [
                 `ui-calendar--color-${this.color}`,
-                `ui-calendar--orientation-${this.orientation}`
+                { 'is-raised': this.raised }
             ];
-        },
-
-        headerYear() {
-            return this.value ? this.value.getFullYear() : this.today.getFullYear();
-        },
-
-        headerDay() {
-            return this.value ?
-                dateUtils.getDayAbbreviated(this.value, this.lang) :
-                dateUtils.getDayAbbreviated(this.today, this.lang);
-        },
-
-        headerDate() {
-            const date = this.value ? this.value : this.today;
-
-            return dateUtils.getMonthAbbreviated(date, this.lang) + ' ' +
-                dateUtils.getDayOfMonth(date, this.lang);
         }
     },
 
@@ -163,82 +106,22 @@ export default {
             if (this.value) {
                 this.dateInView = dateUtils.clone(this.value);
             }
-        },
-
-        showYearPicker() {
-            if (this.showYearPicker) {
-                this.$nextTick(() => {
-                    const el = this.$refs.years.querySelector('.is-selected') ||
-                    this.$refs.years.querySelector('.is-current-year');
-
-                    scrollIntoView(el, { marginTop: 126 });
-                });
-            }
         }
     },
 
     methods: {
-        selectYear(year) {
-            const newDate = dateUtils.clone(this.dateInView);
-            newDate.setFullYear(year);
-
-            this.dateInView = this.getDateInRange(newDate);
-            this.showYearPicker = false;
-        },
-
-        getDateInRange(date, fallback) {
-            date = date || fallback;
-
-            if (this.minDate && date.getTime() < this.minDate.getTime()) {
-                return this.minDate;
-            }
-
-            if (this.maxDate && date.getTime() > this.maxDate.getTime()) {
-                return this.maxDate;
-            }
-
-            return date;
-        },
-
-        getYearClasses(year) {
-            return {
-                'is-current-year': this.isYearCurrent(year),
-                'is-selected': this.isYearSelected(year)
-            };
-        },
-
-        isYearCurrent(year) {
-            return year === this.today.getFullYear();
-        },
-
-        isYearSelected(year) {
-            return this.value && year === this.value.getFullYear();
-        },
-
-        isYearOutOfRange(year) {
-            if (this.minDate && year < this.minDate.getFullYear()) {
-                return true;
-            }
-
-            if (this.maxDate && year > this.maxDate.getFullYear()) {
-                return true;
-            }
-
-            return false;
-        },
-
         onDateSelect(date) {
             this.$emit('input', date);
             this.$emit('date-select', date);
         },
 
-        onGoToDate(date, options = { isForward: true }) {
-            this.$refs.month.goToDate(date, options);
-        },
-
         onMonthChange(newDate) {
             this.dateInView = newDate;
             this.$emit('month-change', newDate);
+        },
+
+        goToDate(date) {
+            this.$refs.month.goToDate(date);
         }
     },
 
@@ -252,158 +135,37 @@ export default {
 <style lang="scss">
 @import './styles/imports';
 
-$ui-calendar-padding    : rem(8px) !default;
+$ui-calendar-padding: rem(8px) !default;
 
 .ui-calendar {
+    border-radius: 3px;
     color: $primary-text-color;
     font-family: $font-stack;
-}
+    overflow: hidden;
 
-.ui-calendar__header {
-    color: white;
-    line-height: 1;
-    padding: rem(16px);
-}
+    &.is-raised {
+        box-shadow: 0 0 2px rgba(black, 0.12), 0 2px 2px rgba(black, 0.2);
 
-.ui-calendar__header-year,
-.ui-calendar__header-details {
-    cursor: pointer;
-    opacity: 0.75;
-    transition: opacity 0.2s ease;
-
-    &:hover,
-    body[modality="keyboard"] &:focus,
-    &.is-active {
-        opacity: 1;
+        .ui-calendar__body {
+            border: none;
+        }
     }
 
-    body[modality="keyboard"] &:focus {
-        outline: 1px dotted white;
-        outline-offset: 1px;
+    .ui-calendar__header {
+        height: $ui-calendar-controls-height + rem(8px);
+        padding-left: rem(8px);
+        padding-right: rem(8px);
     }
-}
-
-.ui-calendar__header-year {
-    font-size: rem(15px);
-    font-weight: 600;
-    margin-bottom: rem(8px);
-}
-
-.ui-calendar__header-details {
-    font-size: rem(22px)
 }
 
 .ui-calendar__body {
-    height: ($ui-calendar-cell-size * 6) + $ui-calendar-controls-height +
-        $ui-calendar-month-header-height  + ($ui-calendar-padding * 2);
+    border-radius: 0 0 3px 3px;
     overflow: hidden;
-    padding: $ui-calendar-padding;
     position: relative;
-    width: ($ui-calendar-cell-size * 7) + ($ui-calendar-padding * 2);
-}
-
-.ui-calendar__years {
-    height: rem(312px);
-    list-style: none;
-    margin: -$ui-calendar-padding;
-    overflow-y: auto;
-    padding: rem(8px 0);
-}
-
-.ui-calendar__year {
-    align-items: center;
-    cursor: pointer;
-    display: flex;
-    font-size: rem(16px);
-    height: rem(36px);
-    justify-content: center;
-    outline: none;
-
-    &.is-selected {
-        font-size: rem(24px);
-        font-weight: 600;
-        height: rem(40px);
-    }
-}
-
-// ================================================
-// Orientations
-// ================================================
-
-.ui-calendar--orientation-landscape {
-    display: flex;
-
-    .ui-calendar__header {
-        min-width: rem(128px);
-    }
-
-    .ui-calendar__header-date {
-        margin-bottom: rem(12px);
-    }
-
-    .ui-calendar__header-date {
-        display: block;
-        padding-top: rem(4px);
-    }
-}
-
-// ================================================
-// Colors
-// ================================================
-
-.ui-calendar--color-primary {
-    .ui-calendar__header {
-        background-color: $brand-primary-color;
-    }
-
-    .ui-calendar__year {
-        &:hover,
-        body[modality="keyboard"] &:focus {
-            color: $brand-primary-color;
-        }
-
-        &.is-selected {
-            color: $brand-primary-color;
-        }
-    }
-
-    .ui-calendar-week__date {
-        &.is-today {
-            color: $brand-primary-color;
-        }
-
-        &.is-selected,
-        body[modality="keyboard"] &.is-selected {
-            background-color: $brand-primary-color;
-        }
-    }
-}
-
-.ui-calendar--color-accent {
-    .ui-calendar__header {
-        background-color: $brand-accent-color;
-    }
-
-    .ui-calendar__year {
-        &:hover,
-        body[modality="keyboard"] &:focus {
-            color: $brand-accent-color;
-        }
-
-        &.is-selected {
-            color: $brand-accent-color;
-        }
-    }
-
-    .ui-calendar-week__date {
-        &.is-today {
-            color: $brand-accent-color;
-        }
-
-        &.is-selected,
-        body[modality="keyboard"] &.is-selected {
-            background-color: $brand-accent-color;
-        }
-    }
+    width: 100%;
+    padding: rem(8px);
+    padding-top: rem(4px);
+    border: 1px solid #EEE;
+    border-top: 0;
 }
 </style>